refactor: isolate input fields stories
