+++ packages/react-aria-components/stories/NumberField.stories.tsx
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2022 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {Button, Group, Input, Label, NumberField} from 'react-aria-components';
+import React from 'react';
+
+export default {
+  title: 'React Aria Components'
+};
+
+export const NumberFieldExample = () => (
+  <NumberField data-testid="number-field-example" formatOptions={{style: 'currency', currency: 'USD'}}>
+    <Label>Test</Label>
+    <Group style={{display: 'flex'}}>
+      <Button slot="decrement">-</Button>
+      <Input />
+      <Button slot="increment">+</Button>
+    </Group>
+  </NumberField>
+);
