+++ packages/react-aria-components/stories/index.stories.tsx
@@ -11,7 +11,7 @@
  */
 
 import {action} from '@storybook/addon-actions';
-import {Button, Calendar, CalendarCell, CalendarGrid, Cell, Checkbox, Column, ColumnResizer, DateField, DateInput, DatePicker, DateRangePicker, DateSegment, Dialog, DialogTrigger, DropZone, FileTrigger, Group, Header, Heading, Input, Keyboard, Label, Link, ListBox, ListBoxItem, ListBoxProps, Menu, MenuItem, MenuTrigger, Modal, ModalOverlay, NumberField, OverlayArrow, Popover, Radio, RadioGroup, RangeCalendar, ResizableTableContainer, Row, SearchField, Section, Select, SelectValue, Separator, Slider, SliderOutput, SliderThumb, SliderTrack, Switch, Tab, Table, TableBody, TableHeader, TabList, TabPanel, Tabs, TabsProps, Tag, TagGroup, TagList, Text, TextField, TimeField, ToggleButton, Toolbar, Tooltip, TooltipTrigger, useDragAndDrop} from 'react-aria-components';
+import {Button, Calendar, CalendarCell, CalendarGrid, Cell, Checkbox, Column, ColumnResizer, DateInput, DatePicker, DateRangePicker, DateSegment, Dialog, DialogTrigger, DropZone, FileTrigger, Group, Header, Heading, Input, Keyboard, Label, Link, ListBox, ListBoxItem, ListBoxProps, Menu, MenuItem, MenuTrigger, Modal, ModalOverlay, OverlayArrow, Popover, Radio, RadioGroup, RangeCalendar, ResizableTableContainer, Row, SearchField, Section, Select, SelectValue, Separator, Slider, SliderOutput, SliderThumb, SliderTrack, Switch, Tab, Table, TableBody, TableHeader, TabList, TabPanel, Tabs, TabsProps, Tag, TagGroup, TagList, Text, TextField, ToggleButton, Toolbar, Tooltip, TooltipTrigger, useDragAndDrop} from 'react-aria-components';
 import {classNames} from '@react-spectrum/utils';
 import clsx from 'clsx';
 import {FocusRing, isTextDropItem, mergeProps, useButton, useClipboard, useDrag} from 'react-aria';
@@ -224,35 +224,6 @@ export const MenuComplex = () => (
       </Menu>
     </Popover>
   </MenuTrigger>
-);
-
-export const NumberFieldExample = () => (
-  <NumberField data-testid="number-field-example" formatOptions={{style: 'currency', currency: 'USD'}}>
-    <Label>Test</Label>
-    <Group style={{display: 'flex'}}>
-      <Button slot="decrement">-</Button>
-      <Input />
-      <Button slot="increment">+</Button>
-    </Group>
-  </NumberField>
-);
-
-export const DateFieldExample = () => (
-  <DateField data-testid="date-field-example">
-    <Label style={{display: 'block'}}>Date</Label>
-    <DateInput className={styles.field} data-testid2="date-input">
-      {segment => <DateSegment segment={segment} className={clsx(styles.segment, {[styles.placeholder]: segment.isPlaceholder})} />}
-    </DateInput>
-  </DateField>
-);
-
-export const TimeFieldExample = () => (
-  <TimeField data-testid="time-field-example">
-    <Label style={{display: 'block'}}>Time</Label>
-    <DateInput className={styles.field}>
-      {segment => <DateSegment segment={segment} className={clsx(styles.segment, {[styles.placeholder]: segment.isPlaceholder})} />}
-    </DateInput>
-  </TimeField>
 );
 
 export const CalendarExample = () => (
